Adding policies to ensure the authorised persons can delete a record from the favourites

+++ resources/views/components/ui/⚡remove-favourite.blade.php
@@ -14,6 +14,8 @@ new class extends Component
 
     public function remove()
     {
+        $this->authorize('delete', $this->favourite);
+        
         $this->favourite->delete();
         $this->js("window.location.reload()");
     }
